Move sign-petition to containers
