+++ src/routes.js
@@ -6,7 +6,7 @@ import { Config } from './config'
 import { loadSession, trackPage } from './actions/sessionActions'
 import { loadOrganization } from './actions/navActions.js'
 import Home from './containers/home'
-import SignPetition from './pages/sign-petition'
+import SignPetition from './containers/sign-petition'
 import ThanksPage from './pages/thanks'
 import SearchPage from './pages/search'
 import PetitionCreatorDashboard from './containers/petition-creator-dashboard'
